Refactor blog and statistic types, enhance portfolio service, and add blog routes for site user

+++ src/services/blog.service.ts
@@ -2,18 +2,26 @@ import prisma from "@/loaders/prisma";
 import { BlogMetricRepository } from "@/repositories/blog-metric.repository";
 import { BlogRepository } from "@/repositories/blog.repository";
 import { CategoryOnBlogRepository } from "@/repositories/category-on-blog.repository";
+import { SiteUserRepository } from "@/repositories/site-user.repository";
+import type { Identity } from "@/types/base.type";
 import type { BlogFilter, CreateBlog } from "@/types/blog.type";
-import { ThrowForbidden } from "@/utils/exception";
+import {
+  ThrowForbidden,
+  ThrowNotFound,
+  ThrowUnauthorized,
+} from "@/utils/exception";
 import { getPaginationMetadata } from "@/utils/pagination";
 
 export class BlogService {
   private blogRepository: BlogRepository;
   private categoryOnBlogRepository: CategoryOnBlogRepository;
   private blogMetricRepository: BlogMetricRepository;
+  private siteUserRepository: SiteUserRepository;
   constructor() {
     this.blogRepository = new BlogRepository();
     this.categoryOnBlogRepository = new CategoryOnBlogRepository();
     this.blogMetricRepository = new BlogMetricRepository();
+    this.siteUserRepository = new SiteUserRepository();
   }
   public async findAll() {
     return this.blogRepository.findAll();
@@ -34,17 +42,58 @@ export class BlogService {
       metadata: { page, page_count, page_size, current_page },
     };
   }
+  public async paginateBySiteUser(site_user_id: string, filter: BlogFilter) {
+    const count = await this.blogRepository.count(filter, {
+      site_user_id,
+    });
+    const { current_page, page, page_count, page_size } = getPaginationMetadata(
+      filter,
+      count
+    );
+    const blogs = await this.blogRepository.paginateBySiteUserId(
+      site_user_id,
+      filter
+    );
+    return {
+      data: blogs,
+      metadata: { page, page_count, page_size, current_page },
+    };
+  }
 
-  public async create(payload: CreateBlog) {
+  public async paginateBySiteUserApiKey(key: string, filter: BlogFilter) {
+    const siteUser = await this.siteUserRepository.findByApiKey(key);
+    if (!siteUser) return ThrowUnauthorized();
+    const count = await this.blogRepository.count(filter, {
+      site_user_id: siteUser.id,
+    });
+    const { current_page, page, page_count, page_size } = getPaginationMetadata(
+      filter,
+      count
+    );
+    const publishedFilter = {
+      ...filter,
+      published: true as const, // Force TypeScript to recognize this as true
+    };
+    const blogs = await this.blogRepository.paginateBySiteUserId(
+      siteUser.id,
+      publishedFilter
+    );
+    return {
+      data: blogs,
+      metadata: { page, page_count, page_size, current_page },
+    };
+  }
+
+  public async create(identity: Identity, payload: CreateBlog) {
     if (payload.categories) {
       return await prisma.$transaction(async (tx) => {
-        const blog = await this.blogRepository.create(payload, tx);
+        const blog = await this.blogRepository.create(payload, identity, tx);
         for (const category of payload.categories!) {
           await this.categoryOnBlogRepository.create(
             {
               category_id: category,
               blog_id: blog.id,
-              assignedBy: blog.admin_id || blog.site_user_id || "",
+              assignedBy: identity.id,
             },
             tx
           );
@@ -52,10 +101,15 @@ export class BlogService {
         return blog;
       });
     }
-    const blog = await this.blogRepository.create(payload);
+    const blog = await this.blogRepository.create(payload, identity);
     return blog;
   }
-  public async update(id: string, payload: CreateBlog) {
+
+  public async update(id: string, identity: Identity, payload: CreateBlog) {
+    const blog = await this.blogRepository.findById(id);
+    if (!blog) return ThrowNotFound();
+    const owner_id = blog.admin_id || blog.site_user_id;
+    if (identity.id !== owner_id) return ThrowUnauthorized();
     return await prisma.$transaction(async (tx) => {
       await this.categoryOnBlogRepository.deleteByBlogId(id);
       const blog = await this.blogRepository.update(id, payload, tx);
@@ -65,7 +119,7 @@ export class BlogService {
             {
               category_id: category,
               blog_id: blog.id,
-              assignedBy: blog.admin_id || blog.site_user_id || "",
+              assignedBy: identity.id,
             },
             tx
           );
@@ -74,9 +128,25 @@ export class BlogService {
       return blog;
     });
   }
-  public async delete(id: string) {
-    return this.blogRepository.delete(id);
+
+  public async increaseView(slug: string) {
+    const blog = await this.blogRepository.findBySlug(slug);
+    if (!blog) return ThrowForbidden("No Record Found");
+    return await prisma.$transaction(async (tx) => {
+      const metric = await this.blogMetricRepository.findByBlogToday(
+        blog.id,
+        tx
+      );
+      //If not found, then create new note metric
+      if (!metric) {
+        await this.blogMetricRepository.createBlogMetric(blog.id, tx);
+        return blog;
+      }
+      await this.blogMetricRepository.increaseView(metric.id, tx);
+      return blog;
+    });
   }
+
   public async publish(id: string) {
     return this.blogRepository.publish(id);
   }
@@ -84,32 +154,7 @@ export class BlogService {
     return this.blogRepository.unpublish(id);
   }
 
-  public async increaseView(slug: string) {
-    const blog = await this.blogRepository.findBySlug(slug);
-    if (!blog) return ThrowForbidden("No Record Found");
-    return await prisma.$transaction(async (tx) => {
-      await this.blogRepository.increaseView(blog.id, blog.view, tx);
-      const noteMetric = await this.blogMetricRepository.findByBlogToday(
-        blog.id,
-        tx
-      );
-      //If not found, then create new note metric
-      if (!noteMetric) {
-        await this.blogMetricRepository.createNoteMetric(
-          {
-            blog_id: blog.id,
-            view: 1,
-          },
-          tx
-        );
-        return blog;
-      }
-      await this.blogMetricRepository.increaseView(
-        noteMetric.id,
-        Number(noteMetric.view),
-        tx
-      );
-      return blog;
-    });
+  public async delete(id: string) {
+    return this.blogRepository.delete(id);
   }
 }
